+++ src/main/java/com/ketheroth/uncrafter/client/gui/screen/inventory/UncrafterScreen.java
@@ -28,7 +28,7 @@ public class UncrafterScreen extends AbstractContainerScreen<UncrafterContainer>
 		if (this.menu.isInputLocked()) {
 			RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
 			RenderSystem.setShaderTexture(0, GUI);
-			this.blit(poseStack, (this.width - this.imageWidth) / 2 + 51, (this.height - this.imageHeight) / 2 + 51, 176, 0, 3, 3);
+			this.blit(poseStack, this.leftPos + 51, this.topPos + 51, 176, 0, 3, 3);
 		}
 	}
 
@@ -36,7 +36,7 @@ public class UncrafterScreen extends AbstractContainerScreen<UncrafterContainer>
 	protected void renderLabels(PoseStack poseStack, int x, int y) {
 		super.renderLabels(poseStack, x, y);
 		this.font.draw(poseStack, this.title, this.titleLabelX, this.titleLabelY, 4210752);
-		this.font.draw(poseStack, playerInventoryTitle.getContents(), this.inventoryLabelX, this.inventoryLabelY, 4210752);
+		this.font.draw(poseStack, this.playerInventoryTitle.getContents(), this.inventoryLabelX, this.inventoryLabelY, 4210752);
 	}
 
 	@Override
@@ -46,6 +46,12 @@ public class UncrafterScreen extends AbstractContainerScreen<UncrafterContainer>
 		int relX = (this.width - this.imageWidth) / 2;
 		int relY = (this.height - this.imageHeight) / 2;
 		this.blit(poseStack, relX, relY, 0, 0, this.imageWidth, this.imageHeight);
+		boolean[] selected = this.menu.selectedIndexes();
+		for (int i = 0; i < selected.length; i++) {
+			if (selected[i]) {
+				this.blit(poseStack, this.leftPos + 93 + i%3 *18, this.topPos + 17 + i/3*18, 176, 3, 16, 16);
+			}
+		}
 	}
 
 }