+++ src/main/java/com/ketheroth/uncrafter/common/inventory/container/AdvancedUncrafterContainer.java
@@ -1,5 +1,6 @@
 package com.ketheroth.uncrafter.common.inventory.container;
 
+import com.ketheroth.uncrafter.common.blockentity.UncrafterBlockEntity;
 import com.ketheroth.uncrafter.common.config.Configuration;
 import com.ketheroth.uncrafter.core.registry.UncrafterContainerTypes;
 import net.minecraft.MethodsReturnNonnullByDefault;
@@ -9,6 +10,7 @@ import net.minecraft.util.Tuple;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.inventory.ClickType;
 import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
@@ -24,10 +26,11 @@ import java.util.List;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-public class AdvancedUncrafterContainer extends AbstractContainerMenu implements IUncrafterContainer {
+public class AdvancedUncrafterContainer extends AbstractContainerMenu {
 
 	private final BlockPos pos;
 	private final Player player;
+	private final UncrafterBlockEntity uncrafter;
 	private final IItemHandler playerInventory;
 	private final OutputHandler outputItems;
 	private final InputHandler inputItems;
@@ -38,28 +41,22 @@ public class AdvancedUncrafterContainer extends AbstractContainerMenu implements
 		super(UncrafterContainerTypes.ADVANCED_UNCRAFTER_CONTAINER.get(), windowId);
 		this.pos = pos;
 		this.player = player;
+		this.uncrafter = (UncrafterBlockEntity) player.level.getBlockEntity(this.pos);
 		this.playerInventory = new InvWrapper(playerInventory);
-		this.outputItems = new OutputHandler(9, this);
-		this.inputItems = new InputHandler(1, this);
-		this.enchantmentHandler = new EnchantmentHandler(6);
+		this.outputItems = this.uncrafter.getOutput();
+		this.inputItems = this.uncrafter.getInput();
+		this.enchantmentHandler = this.uncrafter.getEnchantmentOutput();
 
 		//layout input inventory
 		addSlot(new SlotItemHandler(this.inputItems, 0, 12, 35) {
 			@Override
 			public boolean mayPlace(@NotNull ItemStack stack) {
-				if (stack.getItem().getRegistryName() == null) {
-					return false;
-				}
-				String name = stack.getItem().getRegistryName().toString();
-				if (Configuration.WHITELIST.get().isEmpty()) {
-					return !Configuration.BLACKLIST.get().contains(name) && !Configuration.IMC_BLACKLIST.contains(name);
-				}
-				return Configuration.WHITELIST.get().contains(name);
+				return Configuration.isValidItem(stack.getItem());
 			}
 
 			@Override
 			public boolean mayPickup(Player playerIn) {
-				return !AdvancedUncrafterContainer.this.isInputLocked() && super.mayPickup(playerIn);
+				return !AdvancedUncrafterContainer.this.uncrafter.isInputLocked() && super.mayPickup(playerIn);
 			}
 		});
 		//layout output inventory
@@ -101,19 +98,25 @@ public class AdvancedUncrafterContainer extends AbstractContainerMenu implements
 	}
 
 	@Override
-	public void removed(Player player) {
-		if (player instanceof ServerPlayer) {
-			ItemStack itemstack = this.inputItems.extractItem(0, 64, false);
-			if (!itemstack.isEmpty()) {
-				if (!this.outputItems.isExtracting() && !this.enchantmentHandler.isExtracting()) {
-					if (player.isAlive() && !((ServerPlayer) player).hasDisconnected()) {
-						player.getInventory().placeItemBackInInventory(itemstack);
-					} else {
-						player.drop(itemstack, false);
+	public void clicked(int index, int quickCraft, ClickType clickType, Player player) {
+		if (1 <= index && index < 10) {
+			int maxExtract = this.uncrafter.isAdvanced() ? Configuration.ADVANCED_EXTRACT_AMOUNT.get() : Configuration.EXTRACT_AMOUNT.get();
+			if (this.outputItems.getStackInSlot(index - 1).isEmpty()) {
+				int n = 0;
+				for (int i = 0; i < 9; i++) {
+					if (this.uncrafter.selected[i]) {
+						n++;
 					}
 				}
+				if (this.uncrafter.selected[index - 1]) {
+					this.uncrafter.selected[index - 1] = false;
+				} else {
+					this.uncrafter.selected[index - 1] = n < maxExtract;
+				}
+				this.uncrafter.setChanged();
 			}
 		}
+		super.clicked(index, quickCraft, clickType, player);
 	}
 
 	@Override
@@ -133,9 +136,15 @@ public class AdvancedUncrafterContainer extends AbstractContainerMenu implements
 					}
 				}
 			} else if (1 <= index && index < 10) {// shift-click in output slots
-				// nothing happens
+				if (!this.moveItemStackTo(slotStack, 10, 46, true)) {
+					return ItemStack.EMPTY;
+				}
+				outputItems.extractItem(index - 1, 1, false);
+				return ItemStack.EMPTY;
 			} else if (10 <= index && index < 16 ) {// shift-click in enchantment slots
-				// nothing happens
+				if (!this.moveItemStackTo(slotStack, 10, 46, true)) {
+					return ItemStack.EMPTY;
+				}
 			} else if (index < 52) {// shift-click in inventory slots
 				if (!slots.get(0).hasItem() || slotStack.is(slots.get(0).getItem().getItem())) {//shift click from inventory
 					if (!this.moveItemStackTo(slotStack, 0, 1, false)) {
@@ -158,44 +167,12 @@ public class AdvancedUncrafterContainer extends AbstractContainerMenu implements
 		return itemstack;
 	}
 
-	@Override
-	public boolean canTakeItemForPickAll(ItemStack stack, Slot slot) {
-		return false;
-	}
-
-	@Override
 	public boolean isInputLocked() {
-		return this.outputItems.isExtracting();
+		return this.uncrafter.isInputLocked();
 	}
 
-	@Override
-	public OutputHandler getOutputHandler() {
-		return this.outputItems;
-	}
-
-	@Override
-	public InputHandler getInputHandler() {
-		return this.inputItems;
-	}
-
-	@Override
-	public EnchantmentHandler getEnchantmentHandler() {
-		return this.enchantmentHandler;
-	}
-
-	@Override
-	public Tuple<Item, List<ItemStack>> getCache() {
-		return this.cache;
-	}
-
-	@Override
-	public RecipeManager getRecipeManager() {
-		return this.player.level.getRecipeManager();
-	}
-
-	@Override
-	public boolean isAdvanced() {
-		return true;
+	public boolean[] selectedIndexes() {
+		return this.uncrafter.selected;
 	}
 
 }