+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 1.18.1-1.1.0
+
+### Added
+- shift-click in output slot
+- block-entity to the uncrafter. You can now use container to auto import/export
+
 ## 1.18.1-1.0.0
 
 ### Added